+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 1.0.3 (2021-08-09)
+1. **Added Readme**
+2. **Simpletools\Db\IndexedFile\File**
+   1. Added iterator to yield `$key`
+   2. Moved custom store class check to `__construct` to lazy load
+   3. Added `flock` to avoid 2 scripts connecting at once
+
 ### 1.0.2 (2021-08-08)
 1. **Examples update**